+++ src/main/java/com/example/bookingsystembackend/controller/LoginController.java
@@ -1,13 +1,19 @@
 package com.example.bookingsystembackend.controller;
 
+import com.example.bookingsystembackend.constants.SecurityConstants;
 import com.example.bookingsystembackend.entity.Customer;
 import com.example.bookingsystembackend.repositories.CustomerRepository;
+import com.example.bookingsystembackend.service.UserService;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,7 +21,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+import java.util.*;
+
+import static org.springframework.security.config.Elements.JWT;
 
 @RestController
 public class LoginController {
@@ -28,6 +38,9 @@ public class LoginController {
 
     @Autowired
     AuthenticationManager authenticationManager;
+
+    @Autowired
+    UserService userService;
 
     @GetMapping("/users")
     public List<Customer> getCustomers() {
@@ -54,17 +67,40 @@ public class LoginController {
         return response;
     }
 
-    @PostMapping("/dologin")
+
+    @GetMapping("/dologin")
     public ResponseEntity<String> doLogin(@RequestBody Customer customer) {
-        Authentication authentication = authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(customer.getEmail(), customer.getPwd()));
-        if(authentication.isAuthenticated()){
-            //return JwtResponseDTO.builder()
-            //        .accessToken(jwtService.GenerateToken(authRequestDTO.getUsername()).build();
-            return ResponseEntity.status(HttpStatus.OK)
-                    .body("logged in");
-        } else {
-            throw new UsernameNotFoundException("invalid user request..!!");
+        System.out.println("customer");
+        System.out.println(customer);
+        UserDetails authentication = userService.loadUserByUsername(customer.getEmail());
+
+        String jwt = "";
+
+        System.out.printf("JWT kaldt");
+        if (null != authentication) {
+            SecretKey key = Keys.hmacShaKeyFor(SecurityConstants.JWT_KEY.getBytes(StandardCharsets.UTF_8));
+            jwt = Jwts.builder().setIssuer("Eazy Bank").setSubject("JWT Token")
+                    .claim("username", authentication.getUsername())
+                    .claim("authorities", populateAuthorities(authentication.getAuthorities()))
+                    .setIssuedAt(new Date())
+                    .setExpiration(new Date((new Date()).getTime() + 30000000))
+                    .signWith(key).compact();
+            //response.setHeader(SecurityConstants.JWT_HEADER, jwt);
+            System.out.println(jwt);
         }
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(jwt);
     }
+
+    private String populateAuthorities(Collection<? extends GrantedAuthority> collection) {
+        Set<String> authoritiesSet = new HashSet<>();
+        for (GrantedAuthority authority : collection) {
+            authoritiesSet.add(authority.getAuthority());
+        }
+        return String.join(",", authoritiesSet);
+    }
+
+
+
 
 }
